feat(app): repoint UI to the shared engine, retire duplicated Wi-Fi/login code
Completes the migration: HomeScreen's connect button, LatchTileService,
and LatchWidget now check Wi-Fi state via LatchAppGraph.platform.wifi
(the shared WifiPlatform) instead of raw ConnectivityManager/WifiManager
calls or the old WiFiConnectionDetector/WiFiStateDetector -- this also
fixes the widget's ConnectAction, which still had the stale
activeNetwork-only check that HomeScreen's had already been fixed for
in an earlier commit this session.

WiFiStatusViewModel.connectionStatus and LatchWidgetUpdater now read
LatchAppGraph.engine.status directly (via EngineStatusBridge's
toLegacyStatus(), which stops being a "temporary" bridge and becomes
the permanent Step/Reason -> display-string mapper, the Android
counterpart to desktop's StatusText.kt) instead of relaying through
Android's own ConnectionStatusManager singleton, which had no readers
left once that happened and is deleted. WiFiStatusViewModel's
authenticatePortal()/toggleConnection() were dead code (no caller --
HomeScreen talks to ForegroundService directly) and are removed rather
than migrated.

ForegroundService's bridge is now scoped to what it's actually still
responsible for: keeping the widget refreshed on every engine status
change (the same side effect ConnectionStatusManager.postStatus() used
to have) and keeping Android's own SessionRepository (session/stats
history -- a different concern from login status, with a genuinely
different LiveConnectionStatus shape than the shared one, deliberately
not touched by this migration) in sync via the isLatched transition.
Its "Latched onto Wi-Fi" toast now goes through the shared
AndroidUserNotifier instead of the now-dead UiNotifier.

Deleted, all confirmed zero remaining callers: WiFiConnectionDetector,
WiFiStateDetector, UiNotifier, AutoLoginManager (Android's copy --
core/wifi/AutoLoginManager.kt is authoritative now), CaptivePortalDetector
(Android's copy -- ditto), ConnectionStatusManager (the object; the
ConnectionStatus type it shared a file with stays, still needed as the
Step/Reason mapper's output type, so the file is renamed to
ConnectionStatus.kt to match).

Verified on a real device across two separate scenarios: Wi-Fi
disabled (HomeScreen correctly opened the Wi-Fi settings panel instead
of calling the service, matching original behavior) and Wi-Fi enabled
(full pipeline traced cleanly through the repointed code, same
DNS-resolution-failure path exercised in the previous commit). Zero
crashes, zero exceptions, process stable throughout. LatchTileService/
LatchWidget got the identical fix pattern already proven correct via
HomeScreen but weren't separately exercised by tapping the actual tile/
widget UI in this session.
